Bugfix: ensuring all host names are ordinary strings.
The bug was that the SysLogHandler would broadcast to logged in users
when the message contains unicode strings. We had unicode strings in
our messages since the client from pysvn returns unicode strings in
the results of client.list in each elements repos_path.

I wonder if paths should be unicode strings... but for now this is a
valid bug fix since we do not have unicode characters in host names.